# T&A 42475: Fixes performance issues in question table