--- a/components/ILIAS/Test/classes/Tables/class.ilTestQuestionBrowserTableGUI.php
+++ b/components/ILIAS/Test/classes/Tables/class.ilTestQuestionBrowserTableGUI.php
@@ -138,8 +138,6 @@ class ilTestQuestionBrowserTableGUI
             $this->ctrl,
             $this->data_factory,
             $question_list,
-            $this->test_obj,
-            $this->tree,
             $this->taxonomy,
             $parent_title
         );