[Mantis: 42475](https://mantis.ilias.de/view.php?id=42475)

There are a couple of performance issues regarding the "Add from Pool" & "Add from other Test" tables. With many records the loading times become very noticeable.

I analyzed the underlying code and believe to have found the possible causes. One problem seems to be that the ordering and limiting of the records happens in PHP instead of in the database. This means that no matter how many records should be displayed to the user (partially filtered) all records get returned from the database and get ordered and limited in PHP which could get quite slow. Another problem is that for every record returned more queries are sent to the database to check whether the question has feedback, hints or taxonomies. This can cause an exponential growth of queries. Also the main query is executed multiple times throughout the execution process. For example once to count all found records (pagination) and once with all found records limited to the page size. Overall this approach results in an overhead of queries.

The idea was to move the filtering, ordering and limiting components directly to the database. The database is much more efficient at doing these tasks. This also simplifies the handling on the PHP side, so the found records can be (more or less) directly passed to the user. Due to the higher complexity of some of the relations between tables, the current query increased in size and so did the overall performance.



> To put the code up to a test I duplicated a question inside a test 4600+ times. In another test I went to the "Add from other Test" table and tested the loading times.
> 
> Without the code it takes about 6.2 sec with no filtering, no sorting and unlimited elements per page.
> With the code it takes about 5 sec with no filtering, no sorting and unlimited elements per page.
> Average of 24% performance increase.
> 
> Without the code it takes about 6.8 sec with some filtering, sorting and unlimited elements per page.
> With the code it takes about 5.8 sec with some filtering, sorting and unlimited elements per page.
> Average of 17% performance increase.
> 
> Please keep in mind that this test is meant to show the extreme test case. The regular user will probably limit the amount of elements significantly. This would bring the loading times well below 1 sec.



If this general idea gets welcomed, may I propose going forward to try to put the filtering, ordering and limiting onto the database side? This should increase performance and lower the code complexity.  Although I do not know, if this would overwhelm the databases for bigger ILIAS Installations, in the long run.

I would be glad to know your opinions on this matter.

A not directly related issue: 
[Mantis: 43093](https://mantis.ilias.de/view.php?id=43093)
As a quick note. If ordering logic would to be moved to the database, the Collation would automatically perform the correct Ordering of German (An some others) special characters.