@matheuszych 2025-01-27
Hello @kergomard,
thank You for your feedback. I believe to have resolved all issues you mentioned. In the `ilAssQuestionList` class I changed a couple of other things as well, since I already have touched most of it. I also found a slight bug when trying to filter by `feedback` or `hints` and fixed it.

Best regards
Matheus