@kergomard 2025-01-27
Hi @matheuszych 

Thank you very much for the update!

I've one question: Why did you move the properties of the `QuestionBrowserTable` back to being `protected` and not `private`? If there is no absolute necessity: Could you please move them back to being private, I don't want nobody to get the idea that this table could be reused in any way never ever.

And, you are not going to like me for this, as you followed my wish to the letter, I just didn't formulate my wish precisely enough, sorry: I still don't like the function [`ilAssQuestionList::getTaxonomyFilterExpressions()`](https://github.com/ILIAS-eLearning/ILIAS/pull/8878/files#diff-8348fef9ba500f473b75a6576bf8310d71df03acedaa247cc62441843c8e4b66R248). The naming is correct now, ...but the function just got too long. I think we should split out the lines 251 to 287 into a function `getFilterByAssignedTaxonomyIdsExpression()` called from within `getTaxonomyFilterExpressions()`. We could then also revert back to early returns and we could get rid of the `$forceBypass`-variable, if I'm not completely dreaming right now this could be written as `if ($taxNodes === []) { continue; }`, right?

Thank you very much again and best,
@kergomard