@matheuszych 2025-01-28
Hello @kergomard
i believe the properties were made protected by accident when merging. I reverted that. The method `ilAssQuestionList::getTaxonomyFilterExpressions()` was indeed a bit to long. It is now split. From what it looks like `$forceBypass` can be replaced.

Everything should now be correct.

Best regards
Matheus