@kergomard 2025-01-28
Thank you very much! Merged and picked to trunk.
I moved the check for `$taxNodes` to before the `foreach` and changed a few variable-names.